cid#1448246 Result is not floating-point
Reviewed-on: https://gerrit.libreoffice.org/75905
Tested-by: Jenkins
Tested-by: Caolán McNamara <caolanm@redhat.com>

+++ canvas/workben/canvasdemo.cxx
@@ -131,7 +131,7 @@ class DemoRenderer
 
         void drawGrid()
         {
-            double d, dIncr = maSize.Width() / 3;
+            long d, dIncr = maSize.Width() / 3;
             for ( d = 0; d <= maSize.Width(); d += dIncr )
                 mxCanvas->drawLine( geometry::RealPoint2D( d, 0 ),
                                     geometry::RealPoint2D( d, maSize.Height() ),
